{Generic,Windows}ProcessHandle: move to own files
